+++ playground/supabase-auth/app/flowglad.ts
@@ -1,9 +1,9 @@
-import { FlowgladServer } from '@flowglad/server';
+import { FlowgladServer } from '@flowglad/nextjs/server';
 import { createClient } from '@/utils/supabase/server';
 
 export const flowgladServer = new FlowgladServer({
   supabaseAuth: {
     client: createClient
   },
-  baseURL: 'https://staging.flowglad.com'
+  baseURL: 'http://localhost:3000'
 });